refactor: moves jsonld mapping function for colletions in a separate mapper file
